Update Final SQL Finished (Fix)
There was a mistake written in line 357. It wasn't linking the book loan table and the library branch table so that's why that line had to be changed.  After Making That change, I realized that the loan records I was looking to get retrieved where branch_name ='Sharpstown' weren't coming up. those records where pointing to the wrong library, thus I cahnged the branch Id to pint to 5 instead of 2 since 5 represents branch 'Sharpstown'.